druteus: refactor into module files (unit.cc split into voice_engine, dsp_chain, etc.)
